+++ NEWS.md
@@ -11,3 +11,7 @@
 * Supported 80 types annotation database download (Exclude BioConductor Package)
 * Supported 181 annotation shortcut using R function and external tools
 * Introduced the 1285 cases B-cell acute lymphoblastic leukemia RNA-seq variants (brvar_v1_core and brvar_v1_extra)
+
+## Issues
+
+* supported 1.10.4.3 data.table in released and 64-bit development version of R